Move core::collector mod.rs -> collector.rs
